Move Tables examples / tests to own folder
