+++ ClosedXML_Tests/ClosedXML_Tests.csproj
@@ -76,6 +76,7 @@
     <Compile Include="Examples\PivotTableTests.cs" />
     <Compile Include="Examples\RowsTests.cs" />
     <Compile Include="Examples\StylesTests.cs" />
+    <Compile Include="Examples\TablesTests.cs" />
     <Compile Include="ExcelDocsComparerTests.cs" />
     <Compile Include="Excel\CalcEngine\InformationTests.cs" />
     <Compile Include="Excel\CalcEngine\LogicalTests.cs" />
@@ -222,7 +223,6 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resource\Examples\Misc\CopyingWorksheets.xlsx" />
-    <EmbeddedResource Include="Resource\Examples\Ranges\UsingTables.xlsx" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resource\Examples\Misc\CopyingRowsAndColumns.xlsx" />
@@ -291,7 +291,8 @@
     <EmbeddedResource Include="Resource\Misc\DuplicateImageNames.xlsx" />
     <EmbeddedResource Include="Resource\Misc\TableHeadersWithLineBreaks.xlsx" />
     <EmbeddedResource Include="Resource\Misc\TableWithNameNull.xlsx" />
-    <EmbeddedResource Include="Resource\Examples\Ranges\ResizingTables.xlsx" />
+    <EmbeddedResource Include="Resource\Examples\Tables\ResizingTables.xlsx" />
+    <EmbeddedResource Include="Resource\Examples\Tables\UsingTables.xlsx" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resource\Images\ImageHandling.png" />
